requirements/requirements-test.txt
-r common.txt
pytest-cov==2.8.1
pytest==5.4.1
sure==1.4.11
responses==0.10.12
pytest-pep8==1.0.6
black
flake8
tox
tox-pyenv
wheel
python-dotenv
